fixed decimal point issue

+++ src/main/java/com/unisys/trans/cps/middleware/models/response/TopLanesResponseDTO.java
@@ -11,7 +11,7 @@ public class TopLanesResponseDTO {
 
     private String valueType;
 
-    private Long value;
+    private float value;
 
     private String unit;
